TrackableCollection class now observes
